requirements.txt
tensorflow==2.12.0
streamlit==1.41.1
pillow==11.1.0
numpy==1.23.5
gdown==5.2.0